Ideas Done

+++ app/Http/Controllers/Auth/SessionsController.php
@@ -12,21 +12,36 @@ class SessionsController extends Controller
     {
         $user = Auth::user();
         $sessions = ActiveSession::query()->where("user_id", $user->id)->get();
+
+        $session = ActiveSession::getCurrentSession();
+
+        $now = time();
+
+        $sessionLogin = strtotime($session->created_at);
+
         return response()->json([
             'result' => true,
+            'canRemove' => ($now - $sessionLogin > 86400),
             'sessions' => $sessions,
         ]);
     }
 
     public function delete($id)
     {
-        $user = Auth::user();
-        $session = ActiveSession::query()->where("user_id", $user->id)->where('id', $id)->first();
-        if ($session instanceof ActiveSession) {
-            $session->delete();
-            return response()->json(['result' => true]);
-        }
 
+        $session = ActiveSession::getCurrentSession();
+
+        $now = time();
+
+        $sessionLogin = strtotime($session->created_at);
+        if ($now - $sessionLogin > 86400) {
+            $user = Auth::user();
+            $session = ActiveSession::query()->where("user_id", $user->id)->where('id', $id)->first();
+            if ($session instanceof ActiveSession) {
+                $session->delete();
+                return response()->json(['result' => true]);
+            }
+        }
         return response()->json(['result' => false]);
     }
 }
